refactor(braket): refactor of Braket adapter structure and tests
