+++ packages/devqubit-braket/tests/conftest.py
@@ -13,6 +13,11 @@ from braket.devices import LocalSimulator
 from devqubit_engine.storage.factory import create_registry, create_store
 
 
+# =============================================================================
+# Infrastructure Fixtures
+# =============================================================================
+
+
 @pytest.fixture
 def tracking_root(tmp_path):
     """Create temporary tracking directory."""
@@ -31,106 +36,43 @@ def registry(tracking_root):
     return create_registry(f"file://{tracking_root}")
 
 
+# =============================================================================
+# Real Device Fixtures
+# =============================================================================
+
+
 @pytest.fixture
 def local_simulator():
-    """Real Braket LocalSimulator."""
+    """Real Braket LocalSimulator for integration tests."""
     return LocalSimulator()
+
+
+# =============================================================================
+# Circuit Fixtures
+# =============================================================================
 
 
 @pytest.fixture
 def bell_circuit():
-    """2-qubit Bell state circuit."""
-    circuit = Circuit()
-    circuit.h(0)
-    circuit.cnot(0, 1)
-    return circuit
+    """2-qubit Bell state circuit (no measurements)."""
+    return Circuit().h(0).cnot(0, 1)
+
+
+@pytest.fixture
+def measured_bell_circuit():
+    """2-qubit Bell state circuit with measurements."""
+    return Circuit().h(0).cnot(0, 1).measure([0, 1])
 
 
 @pytest.fixture
 def ghz_circuit():
     """3-qubit GHZ state circuit."""
-    circuit = Circuit()
-    circuit.h(0)
-    circuit.cnot(0, 1)
-    circuit.cnot(1, 2)
-    return circuit
+    return Circuit().h(0).cnot(0, 1).cnot(1, 2)
 
 
-@pytest.fixture
-def non_clifford_circuit():
-    """Circuit with non-Clifford gates (T gate)."""
-    circuit = Circuit()
-    circuit.h(0)
-    circuit.t(0)
-    circuit.cnot(0, 1)
-    return circuit
-
-
-def build_device_properties(
-    *,
-    qubit_count: int = 2,
-    native_gates: list[str] | None = None,
-    connectivity_graph: dict[str, list[int]] | None = None,
-    fully_connected: bool = False,
-    one_qubit_fidelities: dict[int, list[dict]] | None = None,
-    two_qubit_fidelities: dict[str, list[dict]] | None = None,
-    calibration_time: str = "2025-01-02T03:04:05Z",
-) -> dict:
-    """
-    Build Braket device properties dict.
-
-    Parameters
-    ----------
-    qubit_count : int
-        Number of qubits.
-    native_gates : list of str, optional
-        Native gate set.
-    connectivity_graph : dict, optional
-        Connectivity graph {qubit: [neighbors]}.
-    fully_connected : bool
-        Whether device is fully connected.
-    one_qubit_fidelities : dict, optional
-        {qubit: [{"gateName": str, "fidelity": float}]}.
-    two_qubit_fidelities : dict, optional
-        {edge_key: [{"gateName": str, "fidelity": float}]}.
-    calibration_time : str
-        Calibration timestamp.
-
-    Returns
-    -------
-    dict
-        Properties dict in Braket format.
-    """
-    props: dict = {
-        "service": {"updatedAt": calibration_time},
-        "paradigm": {
-            "qubitCount": qubit_count,
-            "connectivity": {"fullyConnected": fully_connected},
-        },
-    }
-
-    if native_gates is not None:
-        props["paradigm"]["nativeGateSet"] = native_gates
-
-    if connectivity_graph is not None:
-        props["paradigm"]["connectivity"]["connectivityGraph"] = connectivity_graph
-
-    if one_qubit_fidelities or two_qubit_fidelities:
-        props["standardized"] = {}
-
-        if one_qubit_fidelities:
-            props["standardized"]["oneQubitProperties"] = {
-                str(q): {"oneQubitGateFidelity": fids}
-                for q, fids in one_qubit_fidelities.items()
-            }
-
-        if two_qubit_fidelities:
-            props["standardized"]["twoQubitProperties"] = {
-                edge_key: {"twoQubitGateFidelity": fids}
-                for edge_key, fids in two_qubit_fidelities.items()
-            }
-
-    return props
+# =============================================================================
+# Mock Classes
+# =============================================================================
 
 
 class MockProperties:
@@ -143,6 +85,17 @@ class MockProperties:
         return self._d
 
 
+class MockResult:
+    """Mock Braket result for testing."""
+
+    def __init__(self, counts: dict[str, int] | None = None):
+        self._counts = counts or {"00": 50, "11": 50}
+
+    @property
+    def measurement_counts(self) -> dict[str, int]:
+        return self._counts
+
+
 class MockTask:
     """Mock Braket task for testing."""
 
@@ -152,7 +105,7 @@ class MockTask:
         task_id: str = "mock-task-123",
         raise_on_result: Exception | None = None,
     ):
-        self._result_value = result_value
+        self._result_value = result_value or MockResult()
         self._task_id = task_id
         self._raise_on_result = raise_on_result
 
@@ -181,17 +134,6 @@ class MockTaskBatch:
         if self._raise_on_results:
             raise self._raise_on_results
         return self._results_list
-
-
-class MockResult:
-    """Mock Braket result for testing."""
-
-    def __init__(self, counts: dict[str, int] | None = None):
-        self._counts = counts or {"00": 50, "11": 50}
-
-    @property
-    def measurement_counts(self) -> dict[str, int]:
-        return self._counts
 
 
 class MockDevice:
@@ -231,57 +173,33 @@ class MockDevice:
         return "AVAILABLE"
 
     @property
-    def provider_name(self) -> str:
-        return "Amazon Braket"
-
-    @property
     def type(self) -> str:
         return "QPU"
 
     def run(self, task_spec: Any, *args, **kwargs) -> Any:
-        """Record call and return configured value or raise."""
-        self._run_calls.append(
-            {
-                "task_spec": task_spec,
-                "args": args,
-                "kwargs": kwargs,
-            }
-        )
+        self._run_calls.append({"task_spec": task_spec, "args": args, "kwargs": kwargs})
         if self._run_raises:
             raise self._run_raises
-        if self._run_returns is not None:
-            return self._run_returns
-        # Default: return a mock task with mock result
-        return MockTask(result_value=MockResult())
+        return self._run_returns or MockTask()
 
     def run_batch(self, task_specs: list[Any], *args, **kwargs) -> Any:
-        """Record call and return configured value or raise."""
         self._run_batch_calls.append(
-            {
-                "task_specs": task_specs,
-                "args": args,
-                "kwargs": kwargs,
-            }
+            {"task_specs": task_specs, "args": args, "kwargs": kwargs}
         )
         if self._run_batch_raises:
             raise self._run_batch_raises
-        if self._run_batch_returns is not None:
-            return self._run_batch_returns
-        # Default: return a mock batch
-        return MockTaskBatch(results_list=[MockResult() for _ in task_specs])
+        return self._run_batch_returns or MockTaskBatch(
+            [MockResult() for _ in task_specs]
+        )
 
 
 class MockProgramSet:
-    """
-    Mock ProgramSet for testing submission handling.
-
-    ProgramSet should be sent to device.run() as-is, not converted to list.
-    """
+    """Mock ProgramSet for testing submission handling."""
 
     def __init__(self, circuits: list[Any], marker: str = "program_set_marker"):
         self._circuits = circuits
         self._marker = marker
-        self._entries = [_MockProgramSetEntry(c) for c in circuits]
+        self._entries = [type("Entry", (), {"circuit": c})() for c in circuits]
 
     @property
     def entries(self) -> list:
@@ -293,30 +211,89 @@ class MockProgramSet:
 
     @property
     def marker(self) -> str:
-        """Marker to verify object identity in tests."""
         return self._marker
 
     def to_ir(self) -> str:
         return "mock_ir"
 
 
-class _MockProgramSetEntry:
-    """Mock entry in a ProgramSet."""
+# =============================================================================
+# Mock Device Factory
+# =============================================================================
 
-    def __init__(self, circuit: Any):
-        self.circuit = circuit
+
+def _build_device_properties(
+    *,
+    qubit_count: int = 2,
+    native_gates: list[str] | None = None,
+    connectivity_graph: dict[str, list[int]] | None = None,
+    fully_connected: bool = False,
+    one_qubit_fidelities: dict[int, list[dict]] | None = None,
+    two_qubit_fidelities: dict[str, list[dict]] | None = None,
+    calibration_time: str = "2025-01-02T03:04:05Z",
+) -> dict:
+    """Build Braket device properties dict."""
+    props: dict = {
+        "service": {"updatedAt": calibration_time},
+        "paradigm": {
+            "qubitCount": qubit_count,
+            "connectivity": {"fullyConnected": fully_connected},
+        },
+    }
+
+    if native_gates is not None:
+        props["paradigm"]["nativeGateSet"] = native_gates
+
+    if connectivity_graph is not None:
+        props["paradigm"]["connectivity"]["connectivityGraph"] = connectivity_graph
+
+    if one_qubit_fidelities or two_qubit_fidelities:
+        props["standardized"] = {}
+        if one_qubit_fidelities:
+            props["standardized"]["oneQubitProperties"] = {
+                str(q): {"oneQubitGateFidelity": fids}
+                for q, fids in one_qubit_fidelities.items()
+            }
+        if two_qubit_fidelities:
+            props["standardized"]["twoQubitProperties"] = {
+                edge_key: {"twoQubitGateFidelity": fids}
+                for edge_key, fids in two_qubit_fidelities.items()
+            }
+
+    return props
 
 
 @pytest.fixture
-def build_properties():
-    """Factory to build Braket device properties."""
-    return build_device_properties
+def device_factory():
+    """Factory that builds a MockDevice with configurable properties."""
+
+    def _factory(
+        *,
+        name: str = "mock_device",
+        run_returns: Any = None,
+        run_raises: Exception | None = None,
+        run_batch_returns: Any = None,
+        run_batch_raises: Exception | None = None,
+        **kwargs,
+    ) -> MockDevice:
+        props = _build_device_properties(**kwargs)
+        return MockDevice(
+            props,
+            name=name,
+            run_returns=run_returns,
+            run_raises=run_raises,
+            run_batch_returns=run_batch_returns,
+            run_batch_raises=run_batch_raises,
+        )
+
+    return _factory
 
 
 @pytest.fixture
-def qpu_properties(build_properties):
-    """Realistic QPU properties with calibration."""
-    return build_properties(
+def mock_device(device_factory):
+    """Default MockDevice with realistic QPU properties."""
+    return device_factory(
+        name="mock_qpu",
         qubit_count=8,
         native_gates=["x", "rz", "cz"],
         connectivity_graph={
@@ -341,83 +318,6 @@ def qpu_properties(build_properties):
 
 
 @pytest.fixture
-def device_factory(build_properties):
-    """Factory that builds a MockDevice with configurable properties."""
-
-    def _factory(
-        *,
-        name: str = "mock_device",
-        run_returns: Any = None,
-        run_raises: Exception | None = None,
-        run_batch_returns: Any = None,
-        run_batch_raises: Exception | None = None,
-        **kwargs,
-    ) -> MockDevice:
-        props = build_properties(**kwargs)
-        return MockDevice(
-            props,
-            name=name,
-            run_returns=run_returns,
-            run_raises=run_raises,
-            run_batch_returns=run_batch_returns,
-            run_batch_raises=run_batch_raises,
-        )
-
-    return _factory
-
-
-@pytest.fixture
-def mock_device(qpu_properties):
-    """Default MockDevice based on qpu_properties."""
-    return MockDevice(qpu_properties)
-
-
-@pytest.fixture
 def mock_program_set(bell_circuit):
     """Mock ProgramSet with identifiable marker."""
     return MockProgramSet([bell_circuit, bell_circuit], marker="test_program_set")
-
-
-@pytest.fixture
-def mock_task():
-    """Factory for creating mock tasks."""
-
-    def _factory(
-        counts: dict[str, int] | None = None,
-        task_id: str = "mock-task-123",
-        raise_on_result: Exception | None = None,
-    ) -> MockTask:
-        result = MockResult(counts) if counts else MockResult()
-        return MockTask(
-            result_value=result,
-            task_id=task_id,
-            raise_on_result=raise_on_result,
-        )
-
-    return _factory
-
-
-@pytest.fixture
-def mock_result():
-    """Factory for creating mock results."""
-
-    def _factory(counts: dict[str, int] | None = None) -> MockResult:
-        return MockResult(counts)
-
-    return _factory
-
-
-@pytest.fixture
-def mock_task_batch():
-    """Factory for creating mock task batches."""
-
-    def _factory(
-        results_list: list[Any] | None = None,
-        raise_on_results: Exception | None = None,
-    ) -> MockTaskBatch:
-        return MockTaskBatch(
-            results_list=results_list,
-            raise_on_results=raise_on_results,
-        )
-
-    return _factory
